Initial commit: bare-bones SDL & GL(ES) app with native & web targets
Code is from older projects, like https://github.com/xndc/vx and its
unpublished C++ rewrite "Reflection".

The build setup (build.py and CMakeLists.txt) is a continuation of the
one used for VX. The VX build scripts were rewritten into Python, kept
up to date, extended and used for multiple projects. This particular
branch of build.py is sadly missing some features (building for iOS,
Clang support, build profiling, etc.) which I'll port over later.

The modified copy of Loguru is from Reflection. It's a nice library. I
changed it to use stb_snprintf for uniform behaviour across platforms
and patched it to work on Emscripten.

The rest of the code is from codebases unrelated to VX, for now, but
I'll be dumping lots of code from VX in here soon enough.