Ground-work for extracting GameObjects into renderables
Port the RenderList mechanism from VX, with some changes.

I thought we could avoid this and just render from the scene graph, but
that didn't really pan out. This is much simpler.

For efficiency, GameObjects should be modified to keep track of their
renderables so we can avoid pointless updates. But that can wait.