Use C++ RTTI instead of type tags
The current type tag system doesn't let us model is-a relationships. We
don't need to at this point, but probably will soon.

RTTI is terrible but it does fix this one issue, and even if we have to
work around its absurd limitations, it's still better than trying to
implement something like LLVM's RTTI system.