Add EditorCamera controlled by RMB/WASD/QE/Space/Shift
It has one major issue at this point, which is that holding RMB first
thing after the window/page shows up doesn't work properly. The cursor
is captured, but we don't get any movement. It works fine after pressing
any of the camera input keys, but not after pressing other keys, so it
seems unlikely to be SDL's fault. Probably something weird I'm doing in
the camera maths.